Earning Filter issue done

+++ app/src/main/java/com/bahamaeatsdriver/model_class/driver_payments/Earning.kt
@@ -1,6 +1,6 @@
 package com.bahamaeatsdriver.model_class.driver_payments
 
 data class Earning(
-    val total_delivery_fee: Int,
-    val total_tip: Int
+    val total_delivery_fee: Double,
+    val total_tip: Double
 )
